fix: split the models into different files
